Optimised some code, mainly by removing "with".
The code's type-checking performance may have degraded due to various
changes performed to the code base, like the switch from direct
definitions to parametrised greatest fixpoints, and perhaps also the use
of overloaded equational reasoning combinators. Maybe it was the
introduction of the one-sided step function that (in conjunction with an
Agda performance bug) made the code so slow that I decided to perform
this optimisation.